Refactor DQNFEPAgent and update learning parameters
